# blueprint/openapi.yaml
openapi: "3.0.3"
info:
  title: Saukele Wedding Registry API
  description: >
    Pool-funded wedding gift registry platform with kinship-tier gift
    suggestions, multi-currency support, and ACID-safe contributions.
  version: "1.0.0"
  contact:
    name: Saukele Support
    email: support@saukele.kz
servers:
  - url: http://localhost:3000/api
    description: Local development server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token. Obtain via POST /auth/login.
        Include as: Authorization: Bearer <token>
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        code:
          type: string
          nullable: true
          example: VALIDATION_ERROR
        error:
          type: string
          example: Resource not found
        message:
          type: string
          nullable: true
          example: Validation failed
        details:
          type: object
          nullable: true
        requestId:
          type: string
          nullable: true

    # ── Monetary precision-safe scalars ───────────────────────────────────
    MoneyAmount:
      type: string
      pattern: '^\\d+(\\.\\d{1,2})?$'
      example: "15000.00"
      description: Decimal monetary amount represented as string to avoid floating-point rounding.
    ExchangeRate:
      type: string
      pattern: '^\\d+(\\.\\d{1,6})?$'
      example: "497.368421"
      description: Decimal exchange rate represented as string with up to 6 fractional digits.

    # ── Pagination cursor ──────────────────────────────────────────────────
    CursorMeta:
      type: object
      properties:
        nextCursor:
          type: integer
          nullable: true
          example: 42
          description: Pass as ?cursor= in next request. Null means no more pages.
        take:
          type: integer
          example: 20
        hasMore:
          type: boolean
          example: true

    # ── User ──────────────────────────────────────────────────────────────
...
    GiftItemCreate:
      type: object
      required: [title, targetAmountKzt]
      properties:
        title:
          type: string
          example: KitchenAid Stand Mixer
        description:
          type: string
          example: 5-quart artisan stand mixer in Empire Red
        imageUrl:
          type: string
          format: uri
          example: https://cdn.saukele.kz/items/kitchenaid.jpg
        targetAmountKzt:
          allOf:
            - $ref: '#/components/schemas/MoneyAmount'
          description: Must be >= 1000.00 KZT
          example: "189000.00"
        priority:
          type: string
          enum: [HIGH, MEDIUM, LOW]
          default: MEDIUM
    GiftItemResponse:
      type: object
      properties:
        id:
          type: integer
          example: 7
        registryId:
          type: integer
          example: 3
        title:
          type: string
          example: KitchenAid Stand Mixer
        description:
          type: string
          example: 5-quart artisan stand mixer in Empire Red
        imageUrl:
          type: string
          nullable: true
          example: https://cdn.saukele.kz/items/kitchenaid.jpg
        targetAmountKzt:
          allOf:
            - $ref: '#/components/schemas/MoneyAmount'
          example: "189000.00"
        currentAmountKzt:
          allOf:
            - $ref: '#/components/schemas/MoneyAmount'
          example: "75600.00"
        targetAmountEur:
          allOf:
            - $ref: '#/components/schemas/MoneyAmount'
          nullable: true
          example: "380.00"
        exchangeRateAtTime:
          allOf:
            - $ref: '#/components/schemas/ExchangeRate'
          nullable: true
          example: "497.368421"
          description: KZT/EUR rate locked at item creation
        lockedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-01-20T08:00:00.000Z"
        status:
          type: string
          enum: [PENDING, FUNDED, PURCHASED, DELIVERED]
          example: PENDING
        priority:
          type: string
          enum: [HIGH, MEDIUM, LOW]
          example: HIGH
        createdAt:
          type: string
          format: date-time
    GiftItemStatusPatch:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [FUNDED, PURCHASED, DELIVERED]
          example: PURCHASED

    # ── Contribution ──────────────────────────────────────────────────────
...
paths:
...
  /registries/{registryId}/items:
    post:
      tags: [GiftItems]
      summary: Add a gift item to a registry (owner or ADMIN)
      security:
        - BearerAuth: []
      parameters:
        - name: registryId
          in: path
          required: true
          schema:
            type: integer
          example: 3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GiftItemCreate'
      responses:
        '201':
          description: Gift item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftItemResponse'
        '400':
          description: Missing required fields or amount < 1000 KZT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the registry owner and not ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Registry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags: [GiftItems]
      summary: List gift items in a registry (owner or ADMIN; cursor-based pagination)
      security:
        - BearerAuth: []
      parameters:
        - name: registryId
          in: path
          required: true
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: integer
          example: 5
        - name: take
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, FUNDED, PURCHASED, DELIVERED]
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, -createdAt, priority]
            default: priority
      responses:
        '200':
          description: Paginated list of gift items
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/GiftItemResponse'
                  meta:
                    $ref: '#/components/schemas/CursorMeta'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the registry owner and not ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Registry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /registries/{registryId}/items/{itemId}:
    get:
      tags: [GiftItems]
      summary: Get a single gift item (owner or ADMIN)
      security:
        - BearerAuth: []
      parameters:
        - name: registryId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Gift item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftItemResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the registry owner and not ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Gift item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [GiftItems]
      summary: Update a gift item (owner or ADMIN, PENDING status only)
      security:
        - BearerAuth: []
      parameters:
        - name: registryId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GiftItemCreate'
      responses:
        '200':
          description: Gift item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftItemResponse'
        '400':
          description: Invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the registry owner and not ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Gift item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Cannot edit item that is FUNDED, PURCHASED, or DELIVERED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [GiftItems]
      summary: Delete a gift item (owner or ADMIN, PENDING status only)
      security:
        - BearerAuth: []
      parameters:
        - name: registryId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Gift item deleted
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the registry owner and not ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Gift item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Cannot delete item with status FUNDED/PURCHASED/DELIVERED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /registries/{registryId}/items/{itemId}/status:
    patch:
      tags: [GiftItems]
      summary: Advance gift item state machine (owner or ADMIN)
      description: >
        Allowed transitions: PENDING→FUNDED, FUNDED→PURCHASED, PURCHASED→DELIVERED.
        Reverse transitions are forbidden. FUNDED→PURCHASED requires
        currentAmountKzt >= targetAmountKzt.
      security:
        - BearerAuth: []
      parameters:
        - name: registryId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GiftItemStatusPatch'
      responses:
        '200':
          description: Status advanced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftItemResponse'
        '400':
          description: Invalid transition or funding not complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Cannot advance to PURCHASED — item is not fully funded
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the registry owner and not ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Gift item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Status transition not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── CONTRIBUTIONS ─────────────────────────────────────────────────────────
...